BUG: Avoid dependency loop; move libtiff from host
To avoid the dependency loop which causes conda-build to try to or
succeed in installing a previous build of libwebp in PREFIX, libtiff
was removed from all requirements/host sections. This is kind of hacky?
but what it does it allowed libwebp to be built against a libtiff that
requires libwebp without having to clobber the previous libwebp during
install to PREFIX. I tried using build/always_include_files to overwrite
the previous libwebp build, but that didn't work for unknown reasons.

The drawback with this approach is that breaks the ability to
cross-compile because conda-build might not choose the correct libtiff
for linking because it is no longer in the host section.